refactor(api-credentials): isolate telemetry parsers
